Modified code to build as both ANSI and Unicode. Fixed up code that used int for sizes and indexes to use size_t.

+++ Client/NetDDECltSocket.cpp
@@ -99,7 +99,7 @@ void CNetDDECltSocket::ReadResponsePacket(CNetDDEPacket& oPacket, uint nPacketID
 		if (m_aoPackets.Size())
 		{
 			// Find our response.
-			for (int i = 0; i < m_aoPackets.Size(); ++i)
+			for (size_t i = 0; i < m_aoPackets.Size(); ++i)
 			{
 				CNetDDEPacket* pPacket = m_aoPackets[i];
